+++ src/testcontainers_atproto/__init__.py
@@ -7,19 +7,23 @@ from testcontainers_atproto.car import CarBlock, CarFile, parse_car
 from testcontainers_atproto.container import PDSContainer
 from testcontainers_atproto.errors import XrpcError
 from testcontainers_atproto.firehose import FirehoseSubscription
+from testcontainers_atproto.rate_limit import CreateSession, RateLimitTarget
 from testcontainers_atproto.ref import RecordRef
-from testcontainers_atproto.seed import Seed
+from testcontainers_atproto.seed import Seed, seed_from_dict
 from testcontainers_atproto.world import World
 
 __all__ = [
     "Account",
     "CarBlock",
     "CarFile",
+    "CreateSession",
     "FirehoseSubscription",
     "PDSContainer",
+    "RateLimitTarget",
     "RecordRef",
     "Seed",
     "World",
+    "seed_from_dict",
     "XrpcError",
     "parse_car",
 ]